Create llmsCheck.yml (#403)
* Create llmsCheck.yml

CI checks if the llms.txt and llms-full.txt file have been updated with the changes made on the new branch. If the files match it means that files were not updated, but if they don't match if means that the branch updated the files. 
This does not ensure that the changes made to the llm files are correct but ensures that at least the files are being updated.

* Update llmsCheck.yml

changed directory to compare against public version of llm files